backend/requirements.txt
asgiref==3.8.1
Django==5.2.1
django-cors-headers==4.7.0
django-extensions==4.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
psycopg2-binary==2.9.10
pydotplus==2.0.2
PyJWT==2.9.0
pyparsing==3.2.3
python-decouple==3.8
sqlparse==0.5.3
django-cors-headers==4.7.0
xhtml2pdf==0.2.17
